refactor: Fix Code Modularize & data struct
