--- src/utils/getNewId.js
@@ -1,3 +1,0 @@
-export function getNewId() {
-  return Math.random().toString(36).substr(2, 16);
-}
